// ui3/src/generated/bridge/SettingEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SettingVariant } from "./SettingVariant";

export type SettingEntry = { name: string, category: string, description: string, minValue: number, maxValue: number, namedVariants: Array<SettingVariant>, stepSize: number, value: number, default: number, };